add commenting and error handling
